+++ ranking/src/Ranker.cpp
@@ -1,10 +1,24 @@
 #include "Ranker.h"
 
+#include "BM25.h"
 #include "DynamicRanker.h"
+#include "PositionIndex.h"
 #include "StaticRanker.h"
+#include "TermDictionary.h"
 #include "TextPreprocessor.h"
+#include "data/Document.h"
 #include "spdlog/sinks/basic_file_sink.h"
 
+#include <algorithm>
+#include <cctype>
+#include <cmath>
+#include <cstddef>
+#include <cstdint>
+#include <optional>
+#include <string>
+#include <unordered_map>
+#include <utility>
+#include <vector>
 #include <spdlog/spdlog.h>
 
 #define LOGGING 0
@@ -29,10 +43,32 @@ int CountWordOccurrences(const std::string& text, const std::string& word) {
 }
 }  // namespace
 
+std::unordered_map<std::string, uint32_t>
+GetDocumentFrequencies(const TermDictionary& term_dict, const std::vector<std::pair<std::string, int>>& query) {
+    std::unordered_map<std::string, uint32_t> map;
+    for (const auto& [query, multiplicity] : query) {
+        auto it = map.find(query);
+        if (it != map.end()) {
+            continue;
+        }
+
+
+        std::optional<TermDictionary::TermEntry> termEntry = term_dict.lookup(query);
+        if (!termEntry.has_value()) {
+            map[query] = 0;
+        } else {
+            map[query] = termEntry->postings_count;
+        }
+    }
+
+    return map;
+}
+
 uint32_t GetFinalScore(const std::vector<std::pair<std::string, int>>& query,
                        const data::Document& doc,
                        const data::DocInfo& info,
-                       const PositionIndex& position_index) {
+                       const PositionIndex& position_index,
+                       const std::unordered_map<std::string, uint32_t>& termFreq) {
 
     auto logger = spdlog::get("ranker_logger");
     if (!logger) {
@@ -69,20 +105,15 @@ uint32_t GetFinalScore(const std::vector<std::pair<std::string, int>>& query,
     float densityTitle = 0.0F;
     float densityDescription = 0.0F;
 
+    float weightedBM25 = 0.0F;
+
     for (const auto& [term, multiplicity] : query) {
-        std::vector<uint16_t> urlPositions =
-            position_index.getPositions(mithril::TokenNormalizer::decorateToken(term, FieldType::URL), doc.id);
-        std::vector<uint16_t> titlePositions =
-            position_index.getPositions(mithril::TokenNormalizer::decorateToken(term, FieldType::TITLE), doc.id);
-        std::vector<uint16_t> descPositions =
-            position_index.getPositions(mithril::TokenNormalizer::decorateToken(term, FieldType::DESC), doc.id);
         std::vector<uint16_t> bodyPositions = position_index.getPositions(term, doc.id);
 
-        bool termInUrl = urlPositions.size() > 0;
-        bool termInDescription = descPositions.size() > 0;
+        bool termInDescription =
+            position_index.hasPositions(mithril::TokenNormalizer::decorateToken(term, FieldType::DESC), doc.id);
         bool termInBody = bodyPositions.size() > 0;
-
-        termInUrl = doc.url.find(term) != std::string::npos;
+        bool termInUrl = doc.url.find(term) != std::string::npos;
 
         size_t pos = title.find(term);
         bool termInTitle = pos != std::string::npos;
@@ -92,7 +123,7 @@ uint32_t GetFinalScore(const std::vector<std::pair<std::string, int>>& query,
         } else {
             wordsInUrl++;
 
-            size_t urlOccurences = std::max(CountWordOccurrences(doc.url, term) * term.size(), doc.url.size());
+            size_t urlOccurences = std::min(CountWordOccurrences(doc.url, term) * term.size(), doc.url.size());
             densityUrl += (static_cast<float>(urlOccurences) / static_cast<float>(doc.url.size())) *
                           (static_cast<float>(multiplicity) / static_cast<float>(query.size()));
         }
@@ -104,7 +135,7 @@ uint32_t GetFinalScore(const std::vector<std::pair<std::string, int>>& query,
             earliestPosTitle += (1 / static_cast<float>(pos + 1)) *
                                 (static_cast<float>(multiplicity) / static_cast<float>(query.size()));
 
-            int titleOccurences = std::max(CountWordOccurrences(title, term), (int)doc.title.size());
+            int titleOccurences = std::min(CountWordOccurrences(title, term), (int)doc.title.size());
             densityTitle += (static_cast<float>(titleOccurences) / static_cast<float>(doc.title.size())) *
                             (static_cast<float>(multiplicity) / static_cast<float>(query.size()));
         }
@@ -122,7 +153,12 @@ uint32_t GetFinalScore(const std::vector<std::pair<std::string, int>>& query,
             earliestPosBody += (1 / static_cast<float>(bodyPositions[0] + 1)) *
                                (static_cast<float>(multiplicity) / static_cast<float>(query.size()));
         }
+
+        weightedBM25 += static_cast<float>(BM25Lib->ScoreTermForDoc(info, termFreq.at(term), bodyPositions.size())) *
+                        (static_cast<float>(multiplicity) / static_cast<float>(query.size()));
     }
+
+    float orderedTitleScore = std::sqrt(ranking::dynamic::OrderedMatchScore(query, doc.title));
 
     dynamic::RankerFeatures features{
         // Boolean presence flags
@@ -136,6 +172,7 @@ uint32_t GetFinalScore(const std::vector<std::pair<std::string, int>>& query,
         .coverage_percent_query_title = (wordsInTitle / totalTermsSize),
         .coverage_percent_query_description = (wordsInDesc / totalTermsSize),
 
+        .order_sensitive_title = orderedTitleScore,
         // Query Density percentage features
         .density_percent_query_url = densityUrl,
         .density_percent_query_title = densityTitle,
@@ -146,6 +183,7 @@ uint32_t GetFinalScore(const std::vector<std::pair<std::string, int>>& query,
         .earliest_pos_body = earliestPosBody,
 
         // Precomputed scores
+        .bm25 = weightedBM25,
         .static_rank = static_cast<float>(GetUrlStaticRank(doc.url)),
         .pagerank = info.pagerank_score,
     };