+++ CHANGELOG.md
@@ -74,6 +74,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 - add scripts that made us build idx1(/5)
 - query coverage, density + position based ranking (#159)
 - Link server to mithril managers
+- add order sensitive ranking & BM25 (#166)
 
 ### Fixed
 
@@ -125,6 +126,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 - Hooked up distributed backend with frontend web-server (#162)
 - Sending urls and titles over network from backend (#163)
 - add script to make positions sync points from built idx (#167)
+- Strip all HTML from title on frontend (#168)
+- Fix hardcoded path (#171)
+- optimize usage of positions in ranker (#170)
 
 ## [0.1.0] - 2025-02-07